Added config and routes dirs
